+++ src/test/ExSln2/LedBlinker/hal/Led.cs
@@ -12,6 +12,19 @@ public class Led : FinObj
         _gpio = gpio;
     }
 
+    public void toggle_twice()
+    {
+        // to test that we can transpile a method call that doesn't have a parameter
+        toggle();
+        this.toggle(); // to test `this.` method invocations
+    }
+
+    public static void toggle_twice_static(Led led)
+    {
+        led.toggle();
+        led.toggle();
+    }
+
     // Will toggle the state of the LED
     public void toggle()
     {