+++ src/test/ExSln2/c99/gen/hal_Led.c
@@ -11,6 +11,19 @@ void hal_Led_ctor(hal_Led * self, hal_Gpio * gpio)
     self->_gpio = gpio;
 }
 
+void hal_Led_toggle_twice(hal_Led * self)
+{
+    // to test that we can transpile a method call that doesn't have a parameter
+    hal_Led_toggle(self);
+    hal_Led_toggle(self); // to test `this.` method invocations
+}
+
+void hal_Led_toggle_twice_static(hal_Led * led)
+{
+    hal_Led_toggle(led);
+    hal_Led_toggle(led);
+}
+
 // Will toggle the state of the LED
 void hal_Led_toggle(hal_Led * self)
 {