+++ src/test/ExSln2/c99/gen/hal_Led.h
@@ -3,6 +3,7 @@
 
 #include "hal_Gpio.h"
 #include <stdint.h>
+#include "hal_Led.h"
 
 
 
@@ -15,6 +16,10 @@ struct hal_Led
 
 
 void hal_Led_ctor(hal_Led * self, hal_Gpio * gpio);
+
+void hal_Led_toggle_twice(hal_Led * self);
+
+void hal_Led_toggle_twice_static(hal_Led * led);
 
 // Will toggle the state of the LED
 void hal_Led_toggle(hal_Led * self);